Worked on UI consistency & routing between Start & Guess pages. Cannot figure out why the size of the nav bar is different on Index page in comparison to Start & Guess Pages.
I feel as if I am putting too much time in to a Guessing Game when I know the next iteration is to be done in PUG
instead of EJS. The purpose of the Guessing Game is not to develop a fully functional project, but to learn fundamentals
of web dev.
I am putting this amount of effort into it because I am trying to get some key features working in a small scale
application so that I can take the learnings and apply to a more serious project.

I know there is no need for a hamburger menu and nav bar and all these bells and whistles but I want to
prove to myself that I can research and implement these features at any time.
Design/UI is not my strong suit but I hope to cultivate these skills as I go.
The time is coming to hang this iteration of Guessing Game up.
Tomorrow I will implement the History page, attempt to resolve this nav bar issue, and implement
database integration. I am already connecting to MondoDB Atlas, but I am not writing anything to it.