Fixed Media Query error where zoom was automatically happening. A result of some copy-pasta without really understanding what each piece does. I still don't think I have a complete solution/understanding of Media Query. At first I was missing the meta tag with the viewport attribute. After I had this tag I also had attribute initial-scale added to it which was causing an issue where the zoom was too deep. This was because of that initial-scale attribute. Documentation said by setting it to 1 the zoom should be fine, but it seems like my application required zooming a little farther out. I am not sure why my application does that. Could it be some styles I have set? That is my guess but I am not really sure.
I will be moving forward because it is not that important for a simple Guessing Game application.
Part of me has this perfectionism thing going where I want to leave this app nearly perfect since it is
not the most complex. The other part of my really wants to move on and learn more things.

Final decision will be to move on and learn more.
In regard to mobile first development. The UI is not too complex regardless so I could have done some
more enhancement but I will not continue to refactor.
I will take my learnings apply it to the next iteration of Guessing Game without really adding any
new features. I truly wish to move on from Guessing Game.